mini_vlm/requirements.txt
Byaldi==0.0.7
colpali_engine==0.3.10
datasets==3.2.0
deepspeed==0.15.4
huggingface_hub==0.30.2
ipython==8.12.3
langchain==0.3.25
langchain_deepseek==0.1.3
pandas==2.2.3
pdf2image==1.17.0
peft==0.14.0
Pillow==11.2.1
qwen_vl_utils==0.0.11
rerankers==0.9.1.post1
safetensors==0.5.3
# torch==2.5.1+cu118  # 单独安装
tqdm==4.67.0
transformers==4.51.3
trl==0.11.4
